# config/NMDAOpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API CORONA FINANCIERO - ADMISION",
    "description": "API Corona Financiero Admision",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [{
    "url": "http://localhost:8080/admision/api",
    "description": "Generated server url"
  }],
...
  "paths": {
    "/regla-negocio/actualizar": {
      "put": {
        "tags": ["Reglas de Negocio"],
        "operationId": "actualizar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReglaNegocioCompletoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/perfilamiento/rol": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarRoles",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionRolDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Perfilamiento"],
        "operationId": "actualizarRol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarRolRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Perfilamiento"],
        "operationId": "crearRol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearRolRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/solicitud/": {
      "post": {
        "tags": ["Solicitud Admision"],
        "operationId": "listaSolicitudAdmision",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiltroSolicitudAdmisionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionSolicitudAdmisionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/regla-negocio/registrar": {
      "post": {
        "tags": ["Reglas de Negocio"],
        "operationId": "registrar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReglaNegocioBasicoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/sucursales": {
      "get": {
        "tags": ["Sucursales"],
        "operationId": "listarSucursales",
        "parameters": [{
          "name": "zona",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SucursalDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/solicitud/{prospectoId}": {
      "get": {
        "tags": ["Solicitud Admision"],
        "operationId": "obtenerEstadoSolicitudAdmision",
        "parameters": [{
          "name": "prospectoId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitudAdmisionResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/solicitud/detalle/{solicitudAdmisionId}/{rutProspecto}": {
      "get": {
        "tags": ["Solicitud Admision"],
        "operationId": "detalleSolicitudAdmision",
        "parameters": [{
          "name": "solicitudAdmisionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }, {
          "name": "rutProspecto",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleSolicitudAdmisionResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/regla-negocio/recuperarPorId/{id}": {
      "get": {
        "tags": ["Reglas de Negocio"],
        "operationId": "recuperarPorId",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaNegocioCompletoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/regla-negocio/listarPaginado": {
      "get": {
        "tags": ["Reglas de Negocio"],
        "operationId": "listarPaginado",
        "parameters": [{
          "name": "numPagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "campoOrdena",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": "descripcion"
          }
        }, {
          "name": "direccionOrdena",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": "ASC"
          }
        }, {
          "name": "valorFiltro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionReglaNegocioDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/perfilamiento/rol/{rolId}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "obtenerRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RolDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Perfilamiento"],
        "operationId": "eliminarRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/regla-negocio/eliminarPorId/{id}": {
      "delete": {
        "tags": ["Reglas de Negocio"],
        "operationId": "eliminarPorId",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ReglaNegocioCompletoDTO": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "valor": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "ActualizarRolRequestDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "modulo": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
...
      "FiltroSolicitudAdmisionDTO": {
        "type": "object",
        "properties": {
          "pagina": {
            "type": "integer",
            "format": "int32"
          },
          "rut": {
            "type": "string"
          },
          "estadoEvaluacion": {
            "type": "integer",
            "format": "int64"
          },
          "faseEvaluacion": {
            "type": "integer",
            "format": "int64"
          },
          "fechaDesde": {
            "type": "string",
            "format": "date-time"
          },
          "fechaHasta": {
            "type": "string",
            "format": "date-time"
          },
          "canalAtencion": {
            "type": "integer",
            "format": "int64"
          },
          "sucursal": {
            "type": "integer",
            "format": "int64"
          },
          "zonaGeografica": {
            "type": "string"
          },
          "columnaOrden": {
            "type": "integer",
            "format": "int32"
          },
          "orden": {
            "type": "string",
            "enum": ["ASC", "DESC"]
          }
        }
      },
      "ListaSolicitudAdmisionDTO": {
        "type": "object",
        "properties": {
          "solicitudId": {
            "type": "integer",
            "format": "int64"
          },
          "rut": {
            "type": "string"
          },
          "estadoSolicitud": {
            "type": "string"
          },
          "faseEvaluacion": {
            "type": "string"
          },
          "fechaSolicitud": {
            "type": "string",
            "format": "date"
          },
          "canalAtencion": {
            "type": "string"
          },
          "sucursal": {
            "type": "string"
          },
          "zonaGeograficaSucursal": {
            "type": "string"
          }
        }
      },
      "PaginacionSolicitudAdmisionDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "solicitudes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListaSolicitudAdmisionDTO"
            }
          }
        }
      },
      "ReglaNegocioBasicoDTO": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "valor": {
            "type": "string"
          }
        }
      },
      "CrearRolRequestDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "modulo": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "SucursalDTO": {
        "type": "object",
        "properties": {
          "codigoSucursal": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "zonaGeografica": {
            "type": "string"
          }
        }
      },
...
      "AdmisionFaseResponseDTO": {
        "type": "object",
        "properties": {
          "admisionFaseId": {
            "type": "integer",
            "format": "int64"
          },
          "faseId": {
            "type": "integer",
            "format": "int64"
          },
          "descripcionFase": {
            "type": "string"
          },
          "fechaRegistroFase": {
            "type": "string",
            "format": "date-time"
          },
          "error": {
            "type": "boolean"
          },
          "mensajeInterno": {
            "type": "string"
          },
          "mensajeEjecutivo": {
            "type": "string"
          }
        }
      },
      "SolicitudAdmisionResponseDTO": {
        "type": "object",
        "properties": {
          "solicitudId": {
            "type": "integer",
            "format": "int64"
          },
          "estado": {
            "type": "string"
          },
          "fechaRegistroSolicitud": {
            "type": "string",
            "format": "date-time"
          },
          "fases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdmisionFaseResponseDTO"
            }
          }
        }
      },
      "DetalleEvaluacionComercialResponsDTO": {
        "type": "object",
        "properties": {
          "faseId": {
            "type": "integer",
            "format": "int64"
          },
          "fase": {
            "type": "string"
          },
          "reglaNegocio": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DetalleSolicitudAdmisionResponseDTO": {
        "type": "object",
        "properties": {
          "solicitudId": {
            "type": "integer",
            "format": "int64"
          },
          "fechaSolicitud": {
            "type": "string",
            "format": "date"
          },
          "canal": {
            "type": "string"
          },
          "estadoSolicitud": {
            "type": "string"
          },
          "rut": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "apellidoPaterno": {
            "type": "string"
          },
          "apellidoMaterno": {
            "type": "string"
          },
          "edad": {
            "type": "integer",
            "format": "int32"
          },
          "movil": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "direccion": {
            "$ref": "#/components/schemas/DireccionDTO"
          },
          "ejecutivo": {
            "$ref": "#/components/schemas/FuncionarioDTO"
          },
          "sucursal": {
            "$ref": "#/components/schemas/SucursalDTO"
          },
          "evaluacionComercial": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleEvaluacionComercialResponsDTO"
            }
          }
        }
      },
      "DireccionDTO": {
        "type": "object",
        "properties": {
          "calle": {
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "comuna": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      },
      "FuncionarioDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          }
        }
      },
      "PaginacionReglaNegocioDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "mensajes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReglaNegocioCompletoDTO"
            }
          }
        }
      },
...
      "ListarRolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "PaginacionRolDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListarRolDTO"
            }
          }
        }
      },
      "RolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "moduloId": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}